+++ config/www/nx-displaygrid/services/todo.service.js
@@ -4,6 +4,14 @@
 
 import { TodoListServiceBase } from './todo-list.service.js';
 
+async function callTodoService(hass, domain, service, data, label) {
+    if (typeof hass?.queueCallService === 'function') {
+        await hass.queueCallService(domain, service, data, { label });
+        return;
+    }
+    await hass.callService(domain, service, data);
+}
+
 export class TodoService extends TodoListServiceBase {
     constructor({ debug = false } = {}) {
         super({
@@ -16,6 +24,12 @@ export class TodoService extends TodoListServiceBase {
     async clearCompleted(hass, entityId) {
         if (!hass) throw new Error('Missing hass');
         const resolved = this._resolveEntityId(entityId);
-        await hass.callService('todo', 'remove_completed_items', { entity_id: resolved });
+        await callTodoService(
+            hass,
+            'todo',
+            'remove_completed_items',
+            { entity_id: resolved },
+            `Clear completed in ${resolved}`
+        );
     }
 }